[CHG] Update namespace of shared-tests
